requirements.txt
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.1
certifi==2026.1.4
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.17.0
flask==3.1.0
flask-cors==5.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
lxml==6.0.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
Python-OOXML==0.13
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
